+++ src/senaite/core/browser/static/js/bika.lims.loader.js
@@ -104,9 +104,6 @@ window.bika.lims.controllers =  {
     ".portaltype-worksheet.template-manage_results":
         ['WorksheetManageResultsView'],
 
-    ".portaltype-worksheettemplate.template-base_edit":
-        ['WorksheetTemplateEdit'],
-
     "#worksheet-printview-wrapper":
         ['WorksheetPrintView'],